backend/requirements/base.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
Pillow==12.2.0
python-decouple==3.8
psycopg2-binary==2.9.11
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-ratelimit==4.1.0
drf-spectacular==0.27.1
django-storages==1.14.2
boto3==1.34.14
slugify==0.0.1
python-slugify==8.0.1
Markdown==3.5.1
bleach==6.1.0
